+++ app/controllers/concerns/backend_lookup.rb
@@ -4,7 +4,7 @@
 # Simple mixin to share the backend_lookup method between search controllers
 module BackendLookup
   def backend_lookup
-    (@response, @document_list) = search_results(params)
+    (@response, @document_list) = search_service.search_results
     respond_to do |format|
       format.json do
         @presenter = Blacklight::JsonPresenter.new(@response,
