+++ app/controllers/preview_controller.rb
@@ -1,12 +1,12 @@
 # Renders the preview pane for browse nearby
 class PreviewController < ApplicationController
   include Blacklight::Configurable
-  include Blacklight::SearchHelper
+  include Blacklight::Searchable
   include Blacklight::SearchContext
   copy_blacklight_config_from(CatalogController)
 
   def show
-    @response, @document = fetch(params[:id])
+    @response, @document = search_service.fetch(params[:id])
     respond_to do |format|
       format.html do
         render preview: @document, layout: false
