+++ app/controllers/selected_databases_controller.rb
@@ -1,8 +1,8 @@
 class SelectedDatabasesController < ApplicationController
-  include Blacklight::SearchHelper
+  include Blacklight::Searchable
   def index
     ids = Settings.selected_databases.keys.map(&:to_s)
-    _, solr_documents = fetch(ids, { sort: "title_sort asc", rows: ids.length })
+    _, solr_documents = search_service.fetch(ids, { sort: "title_sort asc", rows: ids.length })
 
     @selected_databases = solr_documents
   end
